pyproject.toml
[project]
name = "testing-pydantic-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic-ai-slim[duckduckgo]>=0.2.11",
    "pydantic-ai[logfire]>=0.2.11",
]
